+++ Sources/DTCollectionViewManager/DTCollectionViewManager.swift
@@ -389,6 +389,7 @@ internal enum EventMethodSignature: String {
     case previewForDismissingContextMenu = "collectionView:previewForDismissingContextMenuWithConfiguration:"
     case willCommitMenuWithAnimator = "collectionView:willCommitMenuWithAnimator:"
     case canEditItemAtIndexPath = "collectionView:canEditItemAtIndexPath:"
+    case selectionFollowsFocusForItemAtIndexPath = "collectionView:selectionFollowsFocusForItemAtIndexPath:"
     
     // UICollectionViewDelegateFlowLayout
     case sizeForItemAtIndexPath = "collectionView:layout:sizeForItemAtIndexPath:"